+++ BaseLib/IndexTreeFile.cpp
@@ -3937,7 +3937,7 @@ void CIndexTreeFile::Print(CChars* pszDest, bool bShowFlags, bool bShowSize)
 	if (bShowSize)
 	{
 		sz.Append("[");
-		sz.Append((unsigned long long int)this->GetUserMemorySize());
+		sz.Append((uint64)this->GetUserMemorySize());
 		sz.Append("] ");
 	}
